+++ app/src/views/explore/index.tsx
@@ -1,71 +1,189 @@
-import { FC, useEffect, useState } from "react";
+import { FC, useEffect, useRef, useState } from "react";
 import personImage from "../../../public/images/person-opens-the-safe-with-the-money.png";
+import { ExplorerCard, ExplorerCardProps } from "../../components/ExplorerCard";
 import CreateGrant from "../../components/CreateGrant";
-import { ExploreCard, ExploreCardProps } from "../../components/ExploreCard";
+import getGrants from "instructions/getGrants";
+import { useWallet } from '@solana/wallet-adapter-react';
+import getProvider from 'instructions/api/getProvider';
+import { notify } from "../../utils/notifications";
+import { BN } from "@project-serum/anchor";
+import getProgram from "../../instructions/api/getProgram";
+import { getProgramInfoPDA } from "../../instructions/pda/getProgramInfoPDA";
+import { PublicKey } from "@solana/web3.js";
+import fetchDataFromArweave from "../../utils/fetchDataFromArweave";
+import fetchGithubUserDataFromUserId from "utils/fetchGithubUserDataFromUserId";
 
-export const ExploreView: FC = ({ }) => {
-  const [projects, setProjects] = useState<ExploreCardProps[]>([]);
+export const ExplorerView: FC = ({ }) => {
+  const [projects, setProjects] = useState<ExplorerCardProps[]>([]);
   const [preview, setpreview] = useState(false);
+  const programInfo = useRef<any>();
+  const currentGrantIndex = useRef(0);
+  const totalGrantsFetched = useRef(0);
+  const [loadingView, setLoadingView] = useState<-1 | 0 | 1>(1); // 1 -> show loading spinner, 0 -> show load more button, -1 -> show none
+
+  const wallet = useWallet()
+
   useEffect(() => {
-    // TODO: fetch projects from the API
-    let exampleProject = {
-      image: "https://api.lorem.space/image/shoes?w=400&h=225",
-      bgColor: "yellow",
-      title: "Minter Project",
-      author: "minter.sol",
-      authorLink: "https://minter.sol",
-      summary:
-        "Make minting process easier with this framework and then do a lot of subsequent lines until we reach more than 3 lines to test for line clamping",
-      projectLink: "/grant/1",
-      amtRaised: 3012.892,
-      numContributors: 76,
-    };
-    setProjects([
-      exampleProject,
-      exampleProject,
-      exampleProject,
-      exampleProject,
-      exampleProject,
-      exampleProject,
-      exampleProject,
-    ]);
-  }, []);
+    fetchGrants()
+  }, [])
+  
+  const fetchGrants = async () => {
+    try {
+      setLoadingView(1);
+      const provider = getProvider(wallet, true);
+  
+      if (!programInfo.current) {
+        const program = getProgram(provider);
+        const programInfoPDA = await getProgramInfoPDA(program);
+        const programInfoFetched = await program.account.programInfo.fetch(programInfoPDA);
+        // console.log(programInfoFetched)
+        if (!programInfoFetched) {
+          setLoadingView(0);
+          return notify({ type: 'error', message: 'error', description: 'Something went wrong! please try again later' });
+        }
+        programInfo.current = programInfoFetched;
+      }
+  
+      const numGrantsToFetchAtATime = 18;
+      let grantsData = [];
+  
+      while (programInfo.current.grantsCount > totalGrantsFetched.current && grantsData.length < numGrantsToFetchAtATime) {
+        const startIndex = currentGrantIndex.current;
+        let endIndex = startIndex + numGrantsToFetchAtATime - grantsData.length - 1;
+        if (endIndex > programInfo.current.grantsCount - 1) {
+          endIndex = programInfo.current.grantsCount - 1;
+        }
+        console.log(programInfo.current.grantsCount, totalGrantsFetched.current);
+  
+        const grants = await getGrants(provider, startIndex, endIndex);
+        console.log(grants);
+        if (grants.err) {
+          setLoadingView(0);
+          return notify({ type: 'error', message: 'error', description: 'Something went wrong! please try again later' });
+        }
+  
+        grantsData = grants.data;
+        totalGrantsFetched.current += grants.data.length;
+  
+        grantsData = grantsData.filter((grant: ExplorerCardProps) => {
+          if (!grant) {
+            return false;
+          }
+  
+          if (grant.dueDate instanceof BN) {
+            grant.dueDate = grant.dueDate.toNumber();
+          }
+          if (new Date().getTime() > grant.dueDate) {
+            return false;
+          }
+  
+          if (grant.lamportsRaised instanceof BN && grant.targetLamports instanceof BN) {
+            grant.lamportsRaised = grant.lamportsRaised.toNumber();
+            grant.targetLamports = grant.targetLamports.toNumber();
+          }
+          if (grant.lamportsRaised >= grant.targetLamports) {
+            return false;
+          }
+  
+          if (!grant.matchingEligible) {
+            return false;
+          }
+  
+          if (grant.isCancelled) {
+            return false;
+          }
+
+          if (grant.author instanceof PublicKey) {
+            grant.author = grant.author.toString();
+          }
+
+          return true;
+        });
+  
+        currentGrantIndex.current += grantsData.length;
+      }
+
+      await Promise.all(grantsData.map(async (grant) => {
+        const arweaveResponse = await fetchDataFromArweave(grant.info);
+        // console.log(arweaveResponse);
+        if (arweaveResponse.err) {
+          return;
+        }
+        const dataFromArweave = arweaveResponse.data;
+        Object.keys(dataFromArweave).map((key) => {
+          grant[key] = dataFromArweave[key];
+        });
+
+        const githubUserDataResponse = await fetchGithubUserDataFromUserId(dataFromArweave.githubUserId);
+        if (githubUserDataResponse.err) {
+          return;
+        }
+        grant.author = githubUserDataResponse.data.name;
+        grant.authorLink = `https://github.com/${githubUserDataResponse.data.login}`;
+        console.log(grant);
+      }));
+  
+      console.log(grantsData);
+      setProjects(grantsData);  
+      if (programInfo.current.grantsCount === totalGrantsFetched.current) {
+        setLoadingView(-1);
+      }
+      else {
+        setLoadingView(0);
+      }
+    } catch (error) {
+      console.log(error);
+      setLoadingView(0);
+      return notify({ type: 'error', message: 'error', description: 'Something went wrong! please try again later' });
+    }
+  }
 
   return (
     <>
-      <div className='p-4 mx-auto md:hero'>
-        <div className='flex-col gap-20 mb-40 hero-content lg:flex-row'>
+      <div className='md:hero mx-auto p-4'>
+        <div className='hero-content flex-col lg:flex-row gap-20 mb-40'>
           <div className='flex flex-col lg:items-start'>
-            <h1 className='mt-20 text-5xl font-bold text-center text-white lg:text-left bg-clip-text'>
+            <h1 className='text-center lg:text-left text-5xl mt-20 font-bold text-white bg-clip-text'>
               Fund Public Goods
               <br />
               Help grow Solana!{" "}
             </h1>
-            <h4 className='my-5 text-xl text-center text-white md:w-full lg:text-left'>
+            <h4 className='md:w-full text-center lg:text-left text-xl text-white my-5'>
               <p>
                 Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
                 <br /> eiusmod tempor incididunt ut labore et dolore magna
                 aliqua.
               </p>
             </h4>
-            <div className='pt-5 text-center'>
-              <a href="#create-grant" className="px-6 py-2 text-sm bg-transparent border rounded-full hover:bg-slate-500 text-fuchsia-300 border-fuchsia-300 btn"  onClick={()=>{ setpreview(true) }}>
+            <div className='text-center pt-5'>
+              <a href="#create-grant" className="bg-transparent hover:bg-slate-500 py-2 px-6 text-fuchsia-300 border border-fuchsia-300 text-sm rounded-full btn"  onClick={()=>{ setpreview(true) }}>
                 CREATE A GRANT
               </a>
               { preview && <CreateGrant setpreview={setpreview} />}
             </div>
           </div>
-          <div className='hidden pt-10 lg:block'>
+          <div className='pt-10 hidden lg:block'>
             <img src={personImage.src} width='420px' />
           </div>
         </div>
       </div>
-      <div className='px-2 mx-auto lg:container'>
+      <div className='mx-auto px-2 lg:container'>
         <div className='flex flex-wrap justify-center gap-8'>
-          {projects.map((props, idx) => (
-            <ExploreCard key={idx} {...props} />
+          {projects.map((props) => (
+            <ExplorerCard {...props} />
           ))}
         </div>
+      </div>
+      <div className="flex justify-center mb-4">
+        {loadingView != -1 && (
+          loadingView == 0 ? (
+            <button className="bg-cyan-300 hover:bg-blue-700 text-black hover:text-white font-bold py-2 px-4 rounded" onClick={fetchGrants}>
+              Load More
+            </button>
+          ) : (
+            <div className='w-12 h-12 rounded-full animate-spin loading-spinner-gradients'></div>
+          )
+        )}
       </div>
     </>
   );
